+++ task/lib/src/widgets/add_task_dialog.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
 
-import 'package:uuid/uuid.dart';
-
-import 'package:taskj/json.dart';
 import 'package:taskw/taskw.dart';
 
 import 'package:task/task.dart';
@@ -33,16 +30,8 @@ class AddTaskDialog extends StatelessWidget {
           onPressed: () async {
             try {
               validateTaskDescription(controller.text);
-              var now = DateTime.now().toUtc();
               StorageWidget.of(context).mergeTask(
-                Task(
-                  (b) => b
-                    ..status = 'pending'
-                    ..uuid = const Uuid().v1()
-                    ..entry = now
-                    ..description = controller.text
-                    ..modified = now,
-                ),
+                taskParser(controller.text),
               );
               Navigator.of(context).pop();
             } on FormatException catch (e, trace) {